fix(reef): move channel lifecycle helper to its own module for lint/deadcode gates
